# pfelf: use mmap to access Prog() and Data()