Can be a fix for https://github.com/open-telemetry/opentelemetry-ebpf-profiler/issues/352.

Use mmap() to access ELF files backed by pfelf.File.

This is the successor to https://github.com/open-telemetry/opentelemetry-ebpf-profiler/pull/450.