@florianl 2025-05-07
Not sure what kind of benchmark is expected. So here is a benchmark from a use case that will benefit from this approach. In this use case we open a Go executable and parse its gopclntab information. This information is needed to extract stack deltas for unwinding but (potentially) also for native Go symbolization. 

```
$ benchstat old.txt new.txt
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/ebpf-profiler/nativeunwind/elfunwindinfo
cpu: 12th Gen Intel(R) Core(TM) i7-12700H
                                          │   old.txt   │               new.txt               │
                                          │   sec/op    │   sec/op     vs base                │
ParseGoPclnTab/PIE_Go_binary-20             616.1µ ± 9%   571.6µ ± 3%   -7.22% (p=0.000 n=10)
ParseGoPclnTab/stripped_PIE_Go_binary-20    315.2µ ± 9%   193.8µ ± 1%  -38.53% (p=0.000 n=10)
ParseGoPclnTab/regular_Go_binary-20         192.1µ ± 2%   132.9µ ± 2%  -30.78% (p=0.000 n=10)
ParseGoPclnTab/regular_ARM64_Go_binary-20   279.0µ ± 5%   106.6µ ± 2%  -61.78% (p=0.000 n=10)
geomean                                     319.4µ        199.1µ       -37.67%

                                          │    old.txt    │                new.txt                │
                                          │     B/op      │     B/op       vs base                │
ParseGoPclnTab/PIE_Go_binary-20             1046.6Ki ± 0%   936.6Ki ±  0%  -10.51% (p=0.000 n=10)
ParseGoPclnTab/stripped_PIE_Go_binary-20    924.28Ki ± 0%   53.09Ki ± 10%  -94.26% (p=0.000 n=10)
ParseGoPclnTab/regular_Go_binary-20         507.88Ki ± 1%   55.09Ki ±  6%  -89.15% (p=0.000 n=10)
ParseGoPclnTab/regular_ARM64_Go_binary-20   450.43Ki ± 0%   10.56Ki ±  0%  -97.66% (p=0.000 n=10)
geomean                                      685.9Ki        73.33Ki        -89.31%

                                          │   old.txt   │               new.txt               │
                                          │  allocs/op  │  allocs/op   vs base                │
ParseGoPclnTab/PIE_Go_binary-20             2.279k ± 0%   2.277k ± 0%   -0.09% (p=0.000 n=10)
ParseGoPclnTab/stripped_PIE_Go_binary-20     17.00 ± 0%    15.00 ± 0%  -11.76% (p=0.000 n=10)
ParseGoPclnTab/regular_Go_binary-20          13.00 ± 0%    12.00 ± 0%   -7.69% (p=0.000 n=10)
ParseGoPclnTab/regular_ARM64_Go_binary-20    13.00 ± 0%    12.00 ± 0%   -7.69% (p=0.000 n=10)
geomean                                      50.58         47.09        -6.90%

```

<details>
<summary> BenchmarkParseGoPclnTab() </summary>

```
func BenchmarkParseGoPclnTab(b *testing.B) {
	tests := map[string]struct {
		elfFile string
	}{
		// helloworld is a very basic Go binary without special build flags.
		"regular Go binary":       {elfFile: "testdata/helloworld"},
		"regular ARM64 Go binary": {elfFile: "testdata/helloworld.arm64"},
		// helloworld.pie is a Go binary that is build with PIE enabled.
		"PIE Go binary": {elfFile: "testdata/helloworld.pie"},
		// helloworld.stripped.pie is a Go binary that is build with PIE enabled and all debug
		// information stripped.
		"stripped PIE Go binary": {elfFile: "testdata/helloworld.stripped.pie"},
	}

	for name, test := range tests {
		name := name
		test := test
		b.Run(name, func(b *testing.B) {
			b.ReportAllocs()
			ef, err := pfelf.Open(test.elfFile)
			require.NoError(b, err)

			ee := elfExtractor{
				file:   ef,
				hooks:  &extractionFilter{},
				deltas: &sdtypes.StackDeltaArray{},
			}
			b.ResetTimer()
			for i := 0; i < b.N; i++ {
				err = ee.parseGoPclntab()
				require.NoError(b, err)
			}
		})
	}
}
```

</details>