@rockdaboot 2025-05-07
Thanks for the benchmark results, @florianl !
That is a great improvement in all 3 categories, more that I anticipated.